requirements.txt
click==8.0.1
Flask==2.0.1
importlib-metadata==4.6.1
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
numpy==1.21.0
pandas==1.3.0
protobuf==3.17.3
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
